implemented dagre to calc x and y position
